Move navbar dir to components, pass links as props
