+++ frontend/src/components/common/layout-components/NavigationBar/NavigationDrawer.tsx
@@ -10,7 +10,10 @@ import ListItem from "@mui/material/ListItem";
 import ListItemButton from "@mui/material/ListItemButton";
 import ListItemText from "@mui/material/ListItemText";
 
-import { type NavigationItem, type NavigationData } from "../navigationdata";
+import {
+  type NavigationItem,
+  type NavigationData
+} from "../../../../navigation-data";
 
 interface Props {
   container: (() => HTMLElement) | undefined;
