+++ frontend/src/components/common/layout-components/NavigationBar/NavigationUserMenu.tsx
@@ -15,7 +15,10 @@ import Tooltip from "@mui/material/Tooltip";
 
 import AccountCircle from "@mui/icons-material/AccountCircle";
 
-import { type NavigationItem, type NavigationData } from "../navigationdata";
+import {
+  type NavigationItem,
+  type NavigationData
+} from "../../../../navigation-data";
 
 interface Props {
   settings: NavigationData;
