+++ frontend/src/layouts/Layout.tsx
@@ -1,13 +1,14 @@
 import React, { type ReactElement } from "react";
 import { Outlet } from "react-router-dom";
 
-import NavigationBar from "./NavigationBar/NavigationBar";
+import NavigationBar from "../components/common/layout-components/NavigationBar/NavigationBar";
+import { navigationItems, settings } from "../navigation-data";
 
 // TODO: Implement proper layouts.
 const Layout = (): ReactElement => {
   return (
     <div>
-      <NavigationBar />
+      <NavigationBar navigationItems={navigationItems} settings={settings} />
       <Outlet />
     </div>
   );
